feat: lec35 JPQL 로 로직 간소화하기

+++ src/main/kotlin/com/group/libraryapp/dto/book/response/BookStatResponse.kt
@@ -4,6 +4,6 @@ import com.group.libraryapp.domain.book.BookType
 
 data class BookStatResponse (
     val type: BookType,
-    var count: Int,
+    var count: Long,
 ) {
 }
